+++ mini_runner.py
@@ -1,0 +1,10 @@
+
+import example.ruin.corner
+import example.ruin.rectangle
+import example.ruin.ruin_corner
+import example.ruin.ruin_rectangle
+import example.ruin.ruin_rectangle_random
+import example.ruin.ruin_three_wall_corner
+import example.ruin.ruin_three_wall_section
+import example.ruin.three_wall_corner
+import example.ruin.three_wall_section