+++ src/cqterrain/floor/__init__.py
@@ -3,3 +3,4 @@ from .RoundBrickFloor import RoundBrickFloor
 from .WoodFloor import WoodFloor
 from .TileGenerator import TileGenerator
 from .ModPattern import ModPattern
+from .TileFloor import TileFloor