+++ src/cqterrain/wall/__init__.py
@@ -1,0 +1,1 @@
+from .BaseWall import BaseWall